Created PDF parser, specific requirement file, json content file, altered readme, altered git ignore
Initial push of the PDF parser which will extract text from the PDFs in a semi organized way and adds the text to the "content" array in the json output file, for now.

Also added "util" and "data" folders. Created a specific requirements txt for the pdf parser which downloads all the necessary packages for the pdf parser to work. Also ignored venv directory and removed from tracking.